Create README.md
Backed by the biggest manufacturer in this industry and two warehouses in CA and PA, True Image is always the first supplier to release new compatible cartridges in the market. We aim to create value for all printer users by providing high quality printer ink cartridges and toner cartridges with best prices.